+++ Restaurant/Business/Services/IngredientService.cs
@@ -20,7 +20,7 @@ namespace Business.Services
 
         public async Task AddAsync(Ingredient model)
         {
-           /* if (model == null)
+            if (model == null)
             {
                 throw new RestaurantException("Model is null");
             }
@@ -28,16 +28,21 @@ namespace Business.Services
             {
                 throw new RestaurantException("Name is empty");
             }
-           */
+          
             await _unitOfWork.IngredientRepository.AddAsync(model);
             await _unitOfWork.SaveAsync();
         }
 
         public async Task DeleteAsync(int modelId)
         {
-            await _unitOfWork.IngredientRepository.DeleteByIdAsync(modelId);
+            var ingredient = await _unitOfWork.IngredientRepository.GetByIdAsync(modelId);
 
-            await _unitOfWork.SaveAsync();
+            if (ingredient != null)
+            {
+                await _unitOfWork.IngredientRepository.DeleteByIdAsync(modelId);
+
+                await _unitOfWork.SaveAsync();
+            }
         }
 
         public async Task<IEnumerable<Ingredient>> GetAllAsync()
